Moved feature and task accordion items to their respective folders
